Clean up GameDriver
